v0.8: Complete SSL Support with PolarSSL.
Now the client manager implements a -S flag that instruct the builder
to use the SSL layer instead of plain sockets. It also takes care of
generate the SSL keys: certificate, RSA key and Diffie Hellman
parameters to improve security. To make use of this feature is required
libpolarssl and openssl utilities.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>