API Level: new support for API Level.
Starting from this patch, DudaC uses the API level number to build
the Stage area, by default it start using API Level 1. It can be
modified using the -V argument, e.g:

 $ ./dudac -V 1 -f -w ...

when a integer number is given, it will be translated to the proper
branch name on Monkey and Duda repos:

 1 => dst-1
 2 => dst-2
 3 => dst-3
 ...

in order to use the development branches it just need the HEAD or
branch name, e.g:

 $ ./dudac -V master -f -w ...

that command will take master branches for both projects.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>