Stack: new Force mode (-F), deprecates fast-run (smart build).
Starting from this patch, Duda Client tags the stage branch version
used and only rebuild the Stage area if it don't exists or if it was
requested through the -F parameter.

This path deprecates the fast-run mode, it will only rebuild the stack
if required, it can be considered a smart build feature.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>