Build: smart build and setup when using SSL
This patch makes the build system to be aware if the stack was
built with SSL or not, also it now makes easy to run a service
with or without SSL without the need to rebuild just switching
the transport layer on fly.

Tested on bare ~/.dudac and DUDA_STACE_PATH environments.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>